+++ rag_quest/worlds/__init__.py
@@ -3,6 +3,7 @@
 from .exporter import WorldExporter
 from .importer import WorldImporter
 from .modules import Module, ModuleManifestError, ModuleRegistry, load_modules
+from .new_module import NewModuleAnswers, run_interactive, write_module
 from .templates import STARTER_WORLDS
 from .validate import ValidationResult, validate_manifest
 
@@ -16,4 +17,7 @@ __all__ = [
     "load_modules",
     "validate_manifest",
     "ValidationResult",
+    "NewModuleAnswers",
+    "write_module",
+    "run_interactive",
 ]